# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: LLM Backend API
  version: 1.0.0
  description: >
    REST API for managing and generating LLM experiment data,
    including health check, experiment storage, mock & real responses,
    and system summaries.

servers:
  - url: http://localhost:4000/api
    description: Local development server
  - url: https://your-production-domain.com/api
    description: Production server

tags:
  - name: Health
    description: Server health and uptime checks
  - name: Summary
    description: High-level aggregated statistics
  - name: Experiments
    description: Experiment storage and retrieval
  - name: Generation
    description: Prompt generation with mock and real responses

paths:
  /health:
    get:
      summary: Health check endpoint
      description: >
        Returns a simple status response to indicate whether the server is running.
      tags:
        - Health
      responses:
        200:
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-10-25T14:32:05.123Z"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Health check failed"

  /summary:
    get:
      summary: Get summary statistics of all experiments
      description: >
        Returns aggregated metrics across all saved experiments,
        including total experiment count, total number of responses,
        and average coherence score.
      tags:
        - Summary
      responses:
        200:
          description: Summary statistics successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalExperiments:
                    type: integer
                    example: 12
                  totalResponses:
                    type: integer
                    example: 180
                  averageCoherence:
                    type: number
                    format: float
                    example: 0.734
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to calculate summary"

  /experiments:
    get:
      summary: Get all experiments
      description: >
        Fetches a list of saved experiments from the store.
        Supports search and pagination using query parameters.
      tags:
        - Experiments
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: false
          description: Text to search for in experiment prompts.
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          required: false
          description: Maximum number of experiments to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          required: false
          description: Number of experiments to skip (for pagination).
      responses:
        200:
          description: A list of experiments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "a5f9b8f1-3b73-4561-9d98-c2c34a89e1e0"
                    prompt:
                      type: string
                      example: "Inception is best"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to load experiments"

  /experiments/{id}:
    get:
      summary: Get experiment details by ID
      description: >
        Fetches a single experiment by its unique ID, including prompt and all generated results.
      tags:
        - Experiments
      parameters:
        - in: path
          name: id
          required: true
          description: The unique ID of the experiment to retrieve.
          schema:
            type: string
            example: "a5f9b8f1-3b73-4561-9d98-c2c34a89e1e0"
      responses:
        200:
          description: Experiment found
          content:
            application/json:
              schema:
                type: object
                properties:
                  experiment:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "a5f9b8f1-3b73-4561-9d98-c2c34a89e1e0"
                      prompt:
                        type: string
                        example: "Inception is best"
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: "cf275134-f5e0-4c5a-9e80-b3289e5adfa7"
                            text:
                              type: string
                              example: "I think you need to do a google search on Inception is best"
                            params:
                              type: object
                              example:
                                temperature: 0.7
                                top_p: 0.9
                                coherence: 0.57
                                frequency_penalty: 0
                            metrics:
                              type: object
                              properties:
                                coherence:
                                  type: number
                                  example: 0.723
                                lexical:
                                  type: number
                                  example: 0.512
                                lengthFit:
                                  type: number
                                  example: 0.083
                                redundancy:
                                  type: number
                                  example: 0.219
                                length:
                                  type: number
                                  example: 64
        404:
          description: Experiment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Experiment not found"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to load experiment"

  /generate:
    post:
      summary: Generate responses for a given prompt
      description: >
        Accepts a prompt and parameter combinations to generate either mock or real LLM responses.
        Returns an experiment ID and generated results.
      tags:
        - Generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  example: "Inception is best"
                combinations:
                  type: array
                  description: List of parameter combinations for generation.
                  items:
                    type: object
                    properties:
                      temperature:
                        type: number
                        example: 0.7
                      top_p:
                        type: number
                        example: 0.9
                      coherence:
                        type: number
                        example: 0.57
                      frequency_penalty:
                        type: number
                        example: 0
                mock:
                  type: boolean
                  default: true
                  example: true
      responses:
        200:
          description: Successfully generated results
          content:
            application/json:
              schema:
                type: object
                properties:
                  experimentId:
                    type: string
                    example: "a63eab8b-5d98-4d8b-9467-9b09c66e3e2c"
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "cf275134-f5e0-4c5a-9e80-b3289e5adfa7"
                        text:
                          type: string
                          example: "I think you need to do a google search on Inception is best"
                        params:
                          type: object
                          description: Parameter combination used for this generation.
                          example:
                            temperature: 0.7
                            top_p: 0.9
                            coherence: 0.57
                            frequency_penalty: 0
                        metrics:
                          type: object
                          properties:
                            coherence:
                              type: number
                              example: 0.723
                            lexical:
                              type: number
                              example: 0.512
                            lengthFit:
                              type: number
                              example: 0.083
                            redundancy:
                              type: number
                              example: 0.219
                            length:
                              type: number
                              example: 64
        400:
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Prompt is required"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to generate responses"